Update Install-AXAgentMSI.ps1
Some changes were made to this script for Quality of life improvements. The main issue we were aiming to resolve is that the current version of this script will fail/exit if the Automox service exists. However,  uninstalling the agent leaves a lingering "service" causing this script to fail even if the agent is not installed.